+++ A3A/addons/core/Templates/Templates/Aegis/Aegis_Civ.sqf
@@ -59,10 +59,12 @@ if (_hasApex) then {
     "C_Van_01_fuel_F", 0.2
     ,"C_Truck_02_fuel_F", 0.1]] call _fnc_saveToTemplate;
 
+["vehiclesCivPlanes", ["Aegis_C_Plane_Transport_01_civil_F" , "C_Plane_Civil_01_racing_F", "C_Plane_Civil_01_F"]] call _fnc_saveToTemplate;
+
 ["vehiclesCivHeli", [
-"C_Heli_Transport_02_civil_F", 
+"C_Heli_Light_02_civil_F", 
 "C_Heli_Light_01_civil_F", 
-"C_Heli_Light_02_civil_F"
+"C_Heli_Transport_02_civil_F"
 ]] call _fnc_saveToTemplate;
 
 //////////////////////////
