+++ A3A/addons/core/Templates/Templates/Aegis/Aegis_Reb_Vehicle_Attributes.sqf
@@ -1,0 +1,6 @@
+["attributesVehicles", [
+    ["I_C_Offroad_02_LMG_F", ["rebCost", 500]],
+    ["I_G_Offroad_01_armor_base_lxWS", ["rebCost", 500]],
+    ["I_G_Offroad_01_armor_armed_lxWS", ["rebCost", 1100]],
+    ["I_G_Offroad_01_armor_AT_lxWS", ["rebCost", 1500]]
+]] call _fnc_saveToTemplate;
